chore: Simplify props passed to token-cell
