+++ ui/components/app/assets/util/formatWithThreshold.ts
@@ -1,0 +1,16 @@
+export const formatWithThreshold = (
+  amount: number | null,
+  threshold: number,
+  locale: string,
+  options: Intl.NumberFormatOptions,
+): string => {
+  if (amount === null) {
+    return '';
+  }
+  if (amount === 0) {
+    return new Intl.NumberFormat(locale, options).format(0);
+  }
+  return amount < threshold
+    ? `<${new Intl.NumberFormat(locale, options).format(threshold)}`
+    : new Intl.NumberFormat(locale, options).format(amount);
+};
